fix (dev_spring_boot) renamed item service classes
